backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
pydantic==2.10.1
pydantic-settings==2.6.1
asyncpg==0.30.0
aiosqlite==0.21.0
alembic==1.14.0
sqlalchemy==2.0.36
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
better-exceptions==0.3.3
openai==1.52.2
google-generativeai==0.8.6
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2
